Changes to TTImageView x. adding a linked image view, which also loads when this guy loads x. removing self from the request.delegates, which was causing memory exceptions x. adding priority, to set the priority of the TTURLRequest
Changes to TTNetwork
x. added priority to TTURLRequest objects, which are used to determine where they go in the waiting queue.
x. making maxConcurrentLoads and flushDelay for request queue as class variables, which can be set custom set
x. also adding a way to add file names to post requests